active song and playlist item all in one component
